Gestion-comercial/sistema_comercial/requirements.txt
flet
langchain
langgraph
langchain-openai
python-dotenv
fpdf2
